+++ src/architecture/hugging_face/retnet.py
@@ -1,0 +1,8 @@
+from torchscale.architecture.retnet import RetNetDecoder
+from models.hugging_face.hugging_face import HuggingFace
+
+class RetNetHF(HuggingFace):
+    def __init__(self, config):
+        self.config = config
+        model = RetNetDecoder(config, embed_tokens=self.make_embeddings(config))
+        super().__init__(model, config)
