+++ src/train_model.py
@@ -9,7 +9,6 @@ import yaml
 from codecarbon import OfflineEmissionsTracker
 from dataset import DataModule
 from datetime import datetime
-from models import RetNetModel, TransformerModel, PerformerModel
 from pathlib import Path
 from pytorch_lightning import Trainer, loggers as pl_loggers
 from pytorch_lightning.callbacks import EarlyStopping, ModelCheckpoint
@@ -18,6 +17,9 @@ from subprocess import run
 from tabulate import tabulate
 from transformers import set_seed
 from torchinfo import summary as model_summary
+from architecture.lightning.performer import PerformerLightning
+from architecture.lightning.retnet import RetNetLightning
+from architecture.lightning.transformer import TransformerLightning
 from utils import Struct
 
 class CustomModelCheckpoint(ModelCheckpoint):
@@ -65,11 +67,11 @@ def train_model(config: Struct):
 
     # Create requested model
     if config.model_type.lower() == "retnet":
-        model = RetNetModel(config)
+        model = RetNetLightning(config)
     elif config.model_type.lower() == "transformer":
-        model = TransformerModel(config)
+        model = TransformerLightning(config)
     elif config.model_type.lower() == "performer":
-        model = PerformerModel(config)
+        model = PerformerLightning(config)
     else:
         raise ValueError(f"Model type '{config.model_type}' not supported!")
